fix: Move fun formatDate to DateFormater.kt
