+++ app/src/main/java/com/london/tudee/presentation/screens/home/HomeScreenUtils.kt
@@ -18,11 +18,3 @@ object HomeScreenUtils {
         return "$day $month $year"
     }
 }
-
-fun formatDate(instant: Instant): String {
-    val date = instant.toLocalDateTime(TimeZone.currentSystemDefault()).date
-    val day = date.dayOfMonth.toString().padStart(2, '0')
-    val month = date.monthNumber.toString().padStart(2, '0')
-    val year = date.year.toString()
-    return "$day-$month-$year"
-}
