+++ app/src/main/java/com/london/tudee/presentation/utils/DateFormater.kt
@@ -1,0 +1,13 @@
+package com.london.tudee.presentation.utils
+
+import kotlinx.datetime.Instant
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.toLocalDateTime
+
+fun formatDate(instant: Instant): String {
+    val date = instant.toLocalDateTime(TimeZone.currentSystemDefault()).date
+    val day = date.dayOfMonth.toString().padStart(2, '0')
+    val month = date.monthNumber.toString().padStart(2, '0')
+    val year = date.year.toString()
+    return "$day-$month-$year"
+}
